+++ MaxMind.MinFraud/Request/Email.cs
@@ -23,12 +23,15 @@ namespace MaxMind.MinFraud.Request
         private static readonly IReadOnlyDictionary<string, string> _typoDomains = new Dictionary<string, string>
         {
             // gmail.com
-            {"35gmai.com", "gmail.com"},
-            {"636gmail.com", "gmail.com"},
+            {"gmai.com", "gmail.com"},
             {"gamil.com", "gmail.com"},
-            {"gmail.comu", "gmail.com"},
+            {"gmali.com", "gmail.com"},
             {"gmial.com", "gmail.com"},
             {"gmil.com", "gmail.com"},
+            {"gmaill.com", "gmail.com"},
+            {"gmailm.com", "gmail.com"},
+            {"gmailo.com", "gmail.com"},
+            {"gmailyhoo.com", "gmail.com"},
             {"yahoogmail.com", "gmail.com"},
             // outlook.com
             {"putlook.com", "outlook.com"},