+++ releasenotes.md
@@ -17,6 +17,9 @@ Release Notes
   their local part when `hashAddress` is used. For example,
   `foo-bar@yahoo.com` will become `foo@yahoo.com` for additional
   `yahoo.com` domains.
+* Duplicate `.com`s are now removed from email domain names when
+  `hashAddress` is used. For example, `example.com.com` will become
+  `example.com`.
 
 4.2.0 (2023-12-05)
 ------------------