+++ releasenotes.md
@@ -28,6 +28,8 @@ Release Notes
 * Additional `gmail.com` domain names with leading digits are now
   normalized when `hashAddress` is used. For example, `100gmail.com` will
   become `gmail.com`.
+* Additional `gmail.com` typos are now normalized when `hashAddress` is
+  used. For example, `gmali.com` will become `gmail.com`.
 
 4.2.0 (2023-12-05)
 ------------------